1:N

+++ articles/views.py
@@ -20,7 +20,7 @@ def detail(request, article_pk):
     # Article.objects.get(pk=article_pk) => 방법1
     article = get_object_or_404(Article, pk=article_pk)
     form = CommentForm()
-    comments = Comment.objects.all()
+    comments = article.comments.all()
     context = {'article': article, 'form': form, 'comments': comments}
     return render(request, 'articles/detail.html', context)
     
@@ -33,8 +33,10 @@ def create(request):
         # Article 을 생성해달라고 하는 요청
         form = ArticleForm(request.POST)
         if form.is_valid():  # 유효성 검사
-            form.save()
-            return redirect('articles:index')
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
+            return redirect('articles:detail', article.pk)
 
     else: # GET
         # Article 을 생성하기 위한 페이지를 달라고 하는 요청
@@ -46,25 +48,32 @@ def create(request):
 @login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            form.save()
-            return redirect('articles:detail', article_pk)
-    else:  # GET method
-        form = ArticleForm(instance=article)
+    if article.user == request.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                form.save()
+                return redirect('articles:detail', article_pk)
+        else:  # GET method
+            form = ArticleForm(instance=article)
+    else:
+        return redirect('articles:detail', article_pk)
     context = {'form': form}
     return render(request, 'articles/update.html', context)
 
 
 @require_POST
 def delete(request, article_pk):
+    
     if request.user.is_authenticated:
 
         # article_pk에 맞는 article 을 꺼낸다.
         # 삭제한다.
         article = get_object_or_404(Article, pk=article_pk)
-        article.delete()
+        if article.user == request.user:
+            article.delete()
+        else:
+            return redirect('articles:detail', article_pk)
     return redirect('articles:index')
 
 
@@ -75,6 +84,7 @@ def create_com(request, article_pk):
         if form.is_valid():
             comment = form.save(commit=False)  # 데이터베이스에는 적용시키지 마라
             comment.article_id = article_pk
+            comment.user = request.user
             form.save()
     return redirect('articles:detail', article_pk)
 
@@ -83,7 +93,8 @@ def create_com(request, article_pk):
 def comments_delete(request, article_pk, comment_pk):
     if request.user.is_authenticated:
         comment = get_object_or_404(Comment, pk=comment_pk)
-        comment.delete()
+        if comment.user == request.user:
+            comment.delete()
         return redirect('articles:detail', article_pk)
     return HttpResponse('You are Unauthorized', status=401)
